+++ sdk/ai/azure-ai-agents/tsp-location.yaml
@@ -1,5 +1,5 @@
 directory: specification/ai-foundry/data-plane/Foundry/src/sdk-agents
-commit: 93de0dc969c929210496bd9dfd77c0d5ff55ebcd
+commit: 2addecf36e7dcca87339cee28b8e51db9c00bdb2
 repo: Azure/azure-rest-api-specs
 additionalDirectories:
   - specification/ai-foundry/data-plane/Foundry/src/agents
